LayoutTests/inspector-obsolete/http-tests/network/cached-resource-destroyed-too-big-discarded-expected.txt

CONSOLE MESSAGE: line 23: Done1.
CONSOLE MESSAGE: line 33: Done2.
Tests cached resource content is discarded when cached resource is destroyed if content size is too big for the resource agent's data storage.

Bug 92108  
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&random=1&size=400
request.type: image
request.content.length after requesting content: 536
Releasing cached resource.
request.content after requesting content: null
